+++ lib/ui/widget/svg/svgs.dart
@@ -2539,4 +2539,10 @@ class SvgIcons {
     width: 17,
     height: 17,
   );
+
+  static const SvgData paymentCard = SvgData(
+    'assets/icons/payment_card.svg',
+    width: 50,
+    height: 50,
+  );
 }